refactor: dataRender
